2015-11-07 21:15 UTC-0300 Fernando Yurisich <fyurisich@oohg.org>    * include/oohg.h    * source/c_activex.c    * source/c_dialogs.c    * source/c_gdiplus.c    * source/h_datepicker.prg    * source/h_media.prg    * source/h_textbox.prg    * source/miniprint.prg      * BCC warnings pacified.    * source/h_ipaddress.prg      * Format.    * source/h_toolbar.prg      - Unneeded code.

+++ source/c_dialogs.c
@@ -1,5 +1,5 @@
 /*
- * $Id: c_dialogs.c,v 1.15 2015-03-09 02:52:06 fyurisich Exp $
+ * $Id: c_dialogs.c,v 1.16 2015-11-08 00:15:38 fyurisich Exp $
  */
 /*
  * ooHG source code:
@@ -158,7 +158,7 @@ HB_FUNC( CHOOSEFONT )
 		 lf.lfStrikeOut = FALSE ;
 	}
 
-	lf.lfCharSet = hb_parni (8) ;
+	lf.lfCharSet = (BYTE) hb_parni( 8 ) ;
 
 	cf.lStructSize = sizeof(CHOOSEFONT);
 	cf.hwndOwner = hwnd ;
